Fixed Bug

+++ app/Http/Controllers/CommentController.php
@@ -23,6 +23,10 @@ class CommentController extends Controller
      */
     public function index(): View|Factory|Application
     {
+        if(!Auth::user()->hasRole('Super Administrator')){
+            abort(403);
+        }
+
         return view('contents.comment.index', [
             'comments' => Comment::all()->transformWith(new CommentTransfermer())->toArray(),
         ]);
@@ -30,6 +34,10 @@ class CommentController extends Controller
 
     public function store(Request $request)
     {
+        if(!Auth::user()->hasRole('Staff')){
+            abort(403);
+        }
+
         Validator::make($request->all(), [
             'contents' => ['required', 'string', 'max:255'],
             'post-slug' => ['required', 'string', 'max:255', 'exists:posts,slug'],
@@ -37,6 +45,11 @@ class CommentController extends Controller
         ])->validate();
 
         $post = Post::findBySlugOrFail($request->input('post-slug'));
+
+        if($post->hasExpried()) {
+            Session::flash('flash_error_message', 'Post has expired');
+            return redirect()->back();
+        }
 
         $comment = new Comment();
         $comment->contents = $request->input('contents');
@@ -58,6 +71,10 @@ class CommentController extends Controller
      */
     public function destroy($id): RedirectResponse
     {
+        if(!Auth::user()->hasRole('Super Administrator')){
+            abort(403);
+        }
+
         Comment::destroy($id);
 
         Session::flash('flash_success_message', 'Comment successfully deleted!');
